+++ src/main/java/com/library/mapper/CategoryMapper.java
@@ -11,7 +11,7 @@ public class CategoryMapper implements IRowMapper<CategoryModel> {
 	public CategoryModel mapRow(ResultSet rs) {
 		CategoryModel category = new CategoryModel();
 		try {
-			category.setId(rs.getInt("id"));
+			category.setId(rs.getLong("id"));
 			category.setCode(rs.getString("code"));
 			category.setName(rs.getString("name"));
 			return category;
